Include Handlebars placeholders in DOM
A room can be created without giving it a name or/and an image. In
order to overcome these situations the following has been
implemented:

Room Name
In case the room hasn't got a name, the app will concatinate the
usernames of the members (excluding the logged in user) and use
the resultant string as the name for the room.

Room Image
In case the room hasn't got an image, the app will use a random
image of a member (excluding the logged in user).