Change query to available rooms
Currently when retrieving the rooms available for the logged in
user we were included a filter to retrieve only those rooms which
the user was a member of. Apart from being redundant, since by the
default the backend was returning the expected rooms, by using this
approach (through the .query() method) EmberJS wasn't updating the
returned result from the .query() method when a new room was
created.