Retrieve all users, including bots
Currently when opening the chat page, we were only retrieving the
non-bot users and this was done so that the bots weren't shown in
the list of users which the logged in user can interact with. In
this commit, this was changed to retrieve all users (incl. bots)
and hide the bots using handlebar logic.

With this approach, apart from moving the logic where it is suppose
to be, we will also get the benefits of EmberJS observables.