Request for messages should only be done for persisted rooms
Currently upon entering a room, the messages of the room were being
requested. This commit changes that so that the messages are only
requested if the room is persisted. On the other hand, if the room
is not (id === null), the id is observed, so that when it changes,
the messages are requested.