List only valid friends
Currently the users with whome the logged in user was able to chat
with (a.k.a Colleague's List) also listed the user himself. This
commit changes this.

Note that the logic to not list bots was moved from the template to
the JS file since it wasn't possible to do multiple checks.