Add attribute in Room resource to determine unread messages
Currently we had no way to determine whether there are unread
messages within a room. This commit includes the uiUnread computed
attribute which goes through the list of unread messages to
deteremine whether there are messages marked as unread in the
room. This attribute observe the current logged in user's list of
unread messages.