Fix issue with unread number
Currently when loading the app, the notification number was not
showing the number of rooms which had unread updates. This was
being caused because unread messages were not being loaded and
therefore the Room IDs of the unread messages were being set to
undefined. This issue was fixed by including the unread messages of
the current logged in user when requesting its resource.