Consider new messages to be read
Currently when a user recieved a new message while inside the room
the message was written in, the app wasn't considering the message
to be read. This was because the only time the messages of a room
where set to be read was when the user opened the room. This commit
changes this so that the messages of a room are set to be read when
the user view the messages themselves.