Hide DOM when no items match the current filter
Currently when no items match the current filter, the filter dialog
showed an empty un ordered list. Due to the padding of the <ul> tag
this was creating an empty space in the dialog. This commit changes
this so that if no items match the current filter string, the <ul>
tag is not rendered.