Fix race condition issue with empty room name
When a room doesn't have a name, the app uses the username of its
members as the room name. However there are times when the room
name is still left empty. After investigating the problem, it was
found that this was happening because there were times when the
rooms where synced before the users, thus when the app tried
loading the users username apart from making an additional HTTP
Request, it was leaving the room name empty.

This was fixed by moving the HTTP Request to retrieve the Rooms
info after the Users request has been retrieved.