Redo filter when params are updated
Currently when changing the room without destroying the invite
dialog instance, the selectable list in the app-select wasn't being
updated. This was simply because the filter wasn't being done. In
fact upon chaning the filter value the selectable list was getting
in sync with the newly selected room.

In order to fix this issue, the app-select was set so that whenever
the provided params are updated, the filtering is done. In addition
to this to invoke this hook in the invite dialog the invited array
is cleared whenever the chat-invite-dialog params are updated (i.e.
the room model).