Re-Include 'index' route
Before the 'index' route was removed it contained logic that checks
whether the user is logged in or not. If he was the user was
redirected to the 'chat' route else he was redirected to the login
page.

After comparing this logic with the 'application' route it was
found that the logic in the route was redundant since the
application was basically doing the same things and therefore the
'index' route was removed.

With this action when the user accessed the 'index' route
(ex. localhost:4200/) he was left there. For this reason it was
decided to reinclude the 'index' route, but its only job was to
redirect the user to the 'chat' route since, authorization was
already being done in the 'application' route.